Abstracted Layout Logic to Separate File Why: The layout logic remains largely similar across the aforementioned three files, hence it's reasonable to abstract the function into a separate module.
How: Created a generateLayout.tsx file under /lib, and the third
parameter handles whether or not to show PEA000 in the course map.
Build tested and complete.

Tags: Refactor, Abstraction, Modularize, Format
